Made changes suggested
