+++ hoi/core/__init__.py
@@ -1,1 +1,2 @@
-
+from .combinatory import combinations  # noqa
+from .it import ent_g, copnorm_1d, copnorm_nd  # noqa
