+++ hoi/metrics/__init__.py
@@ -1,0 +1,1 @@
+from .oinfo_zerolag import oinfo_zerolag  # noqa
